move NextHopResult and its children to a better package
